requirements.txt
numpy
pandas
streamlit
Pillow
xlsxwriter